// web/src/lib/api/generated/model/thesisSpecV2.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Backfade Thesis Compiler
 * OpenAPI spec version: 0.2.0
 */
import type { ThesisAsset } from './thesisAsset';
import type { ThesisReference } from './thesisReference';
import type { ThesisSpecV2ReferenceOrigin } from './thesisSpecV2ReferenceOrigin';

export interface ThesisSpecV2 {
  version: 2;
  narrative: string;
  basket: ThesisAsset[];
  reference: ThesisReference;
  reference_origin: ThesisSpecV2ReferenceOrigin;
}